Fix init issue between Openthread and LCD display Move Lcd files to efr/platform remove unused files
Sync Lighting app with the changes brough to the server session
remove lcd makefile

+++ examples/lighting-app/efr32/src/AppTask.cpp
@@ -132,7 +132,7 @@ int AppTask::Init()
 #ifdef DISPLAY_ENABLED
     chip::SetupPayload payload;
     uint32_t setUpPINCode       = 0;
-    uint32_t setUpDiscriminator = 0;
+    uint16_t setUpDiscriminator = 0;
 
     err = ConfigurationMgr().GetSetupPinCode(setUpPINCode);
     if (err != CHIP_NO_ERROR)
@@ -160,9 +160,8 @@ int AppTask::Init()
         EFR32_LOG("Failed to get Base41 payload for QR code with %s", chip::ErrorStr(err));
     }
 
-    EFR32_LOG("SetupPINCode: [%" PRIu32 "]", setUpPINCode);
+    EFR32_LOG("SetupPINCode: [%09u]", setUpPINCode);
     LCDWriteQRCode((uint8_t *) result.c_str());
-
 #endif
 
     return err;
